Add error handling to the function
